+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the Postman to Insomnia CLI converter will be documented 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.0] - 2025-07-10
+
+### Added
+- **Support for experimental rules** - New `--experimental` flag to make use of experimental pre and post processing rules
+  - These are replacement rules that have not been confirmed by the Insomnia team yet.
+   - The replacements implemented by these rules are suspected to be correct but are not 100% confirmed
+
 ## [1.4.0] - 2025-07-09
 
 ### Added